Fix previous data
This commit fixes some part of our 2021 and 2022 data. We originally thought that some fees were charged on a monthly basis, when they were only charged on a *quarterly* basis. A few banks made that very clear in their 2023 brochure (adding a "par trimestre" here and there), so we're using that knowledge to adjust our previous data.